# openapi/model_backend_info.go
/*
 * vcverifier
 *
 * Backend component to verify credentials
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type BackendInfo struct {

	IssuerDid Did `json:"issuerDid,omitempty"`

	VerifierDid Did `json:"verifierDid,omitempty"`
}